tutorial/requirements.txt
appdirs==1.4.3
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
autocrop==1.3.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.5.3
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==6.7
cmake==3.24.0
cycler==0.11.0
databases==0.6.1
decorator==5.1.1
Django==3.2
django-browser-reload==1.6.0
django-cors-headers==3.13.0
django-heroku==0.3.1
django-jsonfield==1.4.1
django-orm-magic==0.4
djangorestframework==3.13.1
dlib==19.24.0
executing==1.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==0.12
fonttools==4.34.4
fuzzywuzzy==0.15.0
gnewsclient==1.0.2
gunicorn==19.7.1
hackerrank-sdk==1.2.2
html5lib==0.999999999
idna==3.3
imageio==2.21.1
install==1.3.5
ipython==8.5.0
itsdangerous==0.24
jedi==0.18.1
Jinja2==2.9.5
kiwisolver==1.4.4
MarkupSafe==2.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
networkx==2.8.5
numpy==1.23.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
orm==0.3.1
packaging==16.8
panda==0.3.1
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pymessenger==0.0.7.0
pyparsing==2.2.0
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
requests==2.13.0
requests-toolbelt==0.7.1
scikit-image==0.19.3
scipy==1.9.0
six==1.10.0
SQLAlchemy==1.4.41
sqlparse==0.4.2
stack-data==0.5.0
tifffile==2022.8.3
traitlets==5.3.0
typesystem==0.3.1
urllib3==1.26.11
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.12
whitenoise==6.2.0
wit==4.2.0
